requirements.txt
# Google Drive API
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0

# OpenAI API for vision analysis
openai>=1.12.0
httpx>=0.25.0

# Text-to-Speech (simple TTS)
gTTS==2.4.0

# Image processing (for HEIC conversion)
Pillow==10.1.0
pillow-heif==0.13.0

# HTTP requests
requests==2.31.0

